Working to make tests more independent.
Previously the index & dependency repos were already existing while I was doing other things in parallel (i.e. trying to use this project with other repos).

However, to make this fully testable, the dependencies used in the tests should be purpose built for this project's tests and not used elsewhere.